chore: move naive computeMultiPointProof to a test file and use fk20 method
